refactor: removed the duplication of themes
