+++ frontend/src/components/organisms/TemplateCard/index.js
@@ -12,7 +12,7 @@ import getTemplate from "../../../util/template/getTemplate";
 import Template from "../../../util/template/Template";
 import mainLayouts from "../../../util/layouts";
 import mainAnimations from "../../../util/animation";
-import mainThemes from "../../../util/themes";
+import mainThemes from "../../../../../util/themes/themes";
 import mainFonts from "../../../util/fonts";
 
 const TemplateCard = (props) => {
