--- src/animations/README.md
@@ -1,19 +1,0 @@
-## Available animations
-
-You can add animations to your templates without any manual customization.
-
-Use `?animation=ANIMATION` parameter like so :-
-
-```md
-![Quote](https://github-readme-quotes.herokuapp.com/quote?animation=grow_out_in)
-```
-
-### Animations
-
-- Default <br>
-
-![Quote](https://github-readme-quotes.herokuapp.com/quote?theme=dark)
-
-- Grow-out-in <br>
-
-![Quote](https://github-readme-quotes.herokuapp.com/quote?theme=dark&animation=grow_out_in)
